+++ com.io7m.jregions.core/src/main/java/com/io7m/jregions/core/unparameterized/volumes/VolumeFType.java
@@ -65,15 +65,15 @@ public interface VolumeFType extends VolumeValuesFType
   default void checkPreconditions()
   {
     Preconditions.checkPreconditionD(
-      (double) this.maximumX(),
+      this.maximumX(),
       this.maximumX() >= this.minimumX(),
       x -> "X maximum must be >= X minimum");
     Preconditions.checkPreconditionD(
-      (double) this.maximumY(),
+      this.maximumY(),
       this.maximumY() >= this.minimumY(),
       y -> "Y maximum must be >= Y minimum");
     Preconditions.checkPreconditionD(
-      (double) this.maximumZ(),
+      this.maximumZ(),
       this.maximumZ() >= this.minimumZ(),
       z -> "Z maximum must be >= Z minimum");
   }
